Moving journald config files to the system-config role
Added a new variable called remove_rsyslog, which will remove the
rsyslog package and prevent the rsyslog role from doing anything.
If the rsyslog role is included and remove_rsyslog is not defined or
set to false, then the role will set up a remote syslog server just
as it's doing right now.

The journal configuration has been moved to system-config because
systemd-journald should always be configured, regardless of whether
the rsyslog role is part of the playbook or not.